Seprated all the plugins
