--- plugins/openfga-backend/package.json
@@ -1,49 +1,0 @@
-{
-  "name": "@internal/backstage-plugin-openfga-backend",
-  "version": "0.1.0",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
-  "private": true,
-  "publishConfig": {
-    "access": "public",
-    "main": "dist/index.cjs.js",
-    "types": "dist/index.d.ts"
-  },
-  "backstage": {
-    "role": "backend-plugin"
-  },
-  "scripts": {
-    "start": "backstage-cli package start",
-    "build": "backstage-cli package build",
-    "lint": "backstage-cli package lint",
-    "test": "backstage-cli package test",
-    "clean": "backstage-cli package clean",
-    "prepack": "backstage-cli package prepack",
-    "postpack": "backstage-cli package postpack"
-  },
-  "dependencies": {
-    "@backstage/backend-common": "^0.23.2",
-    "@backstage/backend-defaults": "^0.3.3",
-    "@backstage/backend-plugin-api": "^0.6.21",
-    "@backstage/config": "^1.2.0",
-    "@types/express": "*",
-    "express": "^4.17.1",
-    "express-promise-router": "^4.1.0",
-    "winston": "^3.2.1",
-    "node-fetch": "^2.6.7",
-    "yn": "^4.0.0"
-  },
-  "devDependencies": {
-    "@backstage/backend-test-utils": "^0.4.3",
-    "@backstage/cli": "^0.26.7",
-    "@backstage/plugin-auth-backend": "^0.22.8",
-    "@backstage/plugin-auth-backend-module-guest-provider": "^0.1.5",
-    "@types/supertest": "^2.0.12",
-    "supertest": "^6.2.4",
-    "msw": "^1.0.0"
-  },
-  "files": [
-    "dist"
-  ]
-}
