--- plugins/openfga-backend/src/index.ts
@@ -1,2 +1,0 @@
-export * from './service/router';
-export { openfgaPlugin as default } from './plugin';
