--- plugins/openfga-backend/src/plugin.ts
@@ -1,39 +1,0 @@
-import {
-  coreServices,
-  createBackendPlugin,
-} from '@backstage/backend-plugin-api';
-import { createRouter } from './service/router';
-
-/**
- * openfgaPlugin backend plugin
- *
- * @public
- */
-export const openfgaPlugin = createBackendPlugin({
-  pluginId: 'openfga',
-  register(env) {
-    env.registerInit({
-      deps: {
-        httpRouter: coreServices.httpRouter,
-        logger: coreServices.logger,
-        config: coreServices.rootConfig,
-      },
-      async init({
-        httpRouter,
-        logger,
-        config,
-      }) {
-        httpRouter.use(
-          await createRouter({
-            logger,
-            config,
-          }),
-        );
-        httpRouter.addAuthPolicy({
-          path: '/health',
-          allow: 'unauthenticated',
-        });
-      },
-    });
-  },
-});
