--- plugins/openfga-backend/src/service/router.test.ts
@@ -1,30 +1,0 @@
-import { mockServices } from '@backstage/backend-test-utils';
-import express from 'express';
-import request from 'supertest';
-
-import { createRouter } from './router';
-
-describe('createRouter', () => {
-  let app: express.Express;
-
-  beforeAll(async () => {
-    const router = await createRouter({
-      logger: mockServices.logger.mock(),
-      config: mockServices.rootConfig(),
-    });
-    app = express().use(router);
-  });
-
-  beforeEach(() => {
-    jest.resetAllMocks();
-  });
-
-  describe('GET /health', () => {
-    it('returns ok', async () => {
-      const response = await request(app).get('/health');
-
-      expect(response.status).toEqual(200);
-      expect(response.body).toEqual({ status: 'ok' });
-    });
-  });
-});
