--- plugins/openfga-backend/src/setupTests.ts
@@ -1,1 +1,0 @@
-export {};
