+++ plugins/openfga-common/README.md
@@ -1,5 +1,3 @@
-# @internal/backstage-plugin-openfga-common
+# @infosys_ltd/backstage-plugin-openfga-common
 
 Welcome to the common package for the openfga plugin!
-
-_This plugin was created through the Backstage CLI_
