+++ plugins/openfga-common/config.d.ts
@@ -1,0 +1,19 @@
+/**
+ * This file defines the configuration schema for the OpenFGA plugin.
+ */
+export interface Config {
+  openfga?: {
+    /**
+     * @visibility frontend
+     */
+    baseUrl: string;
+    /**
+     * @visibility frontend
+     */
+    storeId: string;
+    /**
+     * @visibility frontend
+     */
+    authorizationModelId: string;
+  };
+}
