+++ plugins/openfga/src/index.ts
@@ -1,6 +1,1 @@
-export { openfgaPlugin, OpenfgaPage } from './plugin';
-export * from './api/openFgaApi';
-export * from './api/OpenFgaClient';
-export * from './api/types';
-export * from './policy';
-export * from './module';
+export * from './plugin';
