+++ plugins/permission-backend-module-catalog-policy/.eslintrc.js
