+++ plugins/permission-backend-module-catalog-policy/README.md
@@ -1,0 +1,5 @@
+# @internal/backstage-plugin-permission-backend-module-catalog-policy
+
+The catalog-policy backend module for the permission plugin.
+
+_This plugin was created through the Backstage CLI_
