+++ plugins/permission-backend-module-catalog-policy/src/index.ts
@@ -1,0 +1,8 @@
+/***/
+/**
+ * The catalog-policy backend module for the permission plugin.
+ *
+ * @packageDocumentation
+ */
+
+export { permissionModuleCatalogPolicy as default } from './module';
